Moved widgets from theme.dart to widgets.dart
